refactor: remove the table-isolation subprocess, whose cause is gone
Table extraction ran in a separate interpreter via
'python -m pdf_mcp._table_worker'. That existed for one reason:
importing pymupdf4llm activated pymupdf.layout and irreversibly
corrupted PyMuPDF's find_tables cell text process-wide, detaching
decimal points so a datasheet's 4.5 came back as 45\n. Extraction
therefore needed an interpreter that had never imported it.

Both halves of the premise are gone. Detection runs on pdfplumber, and
pymupdf4llm is not a dependency at all; the corruption tests had already
degraded to skips because they guarded their own premise.

Measured before removing: the hop cost 0.129s per call on macOS, where
process start is cheap. Extraction of a ruled table goes 0.189s to
0.060s, 3.1x, and the decimal survives (0.25 reads as 0.25), which is
the property the isolation existed to protect. The fast test suite also
drops about 10 seconds.

extract() moves to extractor.extract_tables_for_pages unchanged.
_table_worker.py and parallel.run_module_json are deleted, since nothing
else needed a pristine interpreter. test_table_isolation.py becomes
test_table_cache_version.py, keeping the part that still protects users:
rows cached during the corrupt era hold wrong numbers and must be
dropped rather than served.
